// parallel-web/src/types/ListTasksQuery.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { TaskPriority } from "./TaskPriority";
import type { TaskStatus } from "./TaskStatus";

export type ListTasksQuery = { status: TaskStatus | null, priority: TaskPriority | null, repo_url: string | null, worker_id?: string, search: string | null, created_after?: string, created_before?: string, sort_by: string | null, sort_direction: string | null, cursor: string | null, limit: number | null, offset: number | null, project_id?: string, };